+++ services/gateway/sources/version3/mappings/schema.js
@@ -32,7 +32,9 @@ module.exports = {
 	standardEvent: {
 		schema: '=',
 	},
-	ccm: '=',
+	ccm: {
+		schema: '=',
+	},
 	events: ['data.events', {
 		module: '=,string',
 		name: '=,string',
